mockchain: Move BlockDate stuff to a separate file
